+++ samples/forms.py
@@ -193,7 +193,7 @@ def demo(screen, scene):
     screen.play([Scene([
         Background(screen),
         DemoFrame(screen)
-    ], -1)], stop_on_resize=True, start_scene=scene, allow_int=True)
+    ], -1)], stop_on_resize=True, start_scene=scene)
 
 
 last_scene = None